# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Context-Aware MCP Server",
    "version": "0.1.0"
  },
  "paths": {
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Read Projects",
        "operationId": "read_projects_api_projects_get",
        "parameters": [
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          },
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "title": "Response Read Projects Api Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project Endpoint",
        "operationId": "create_project_endpoint_api_projects_post",
        "parameters": [
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Read Project",
        "operationId": "read_project_api_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/archive": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Archive Project Endpoint",
        "operationId": "archive_project_endpoint_api_projects__project_id__archive_patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectArchiveUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/memberships": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Read Project Memberships",
        "operationId": "read_project_memberships_api_projects__project_id__memberships_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectMembership"
                  },
                  "title": "Response Read Project Memberships Api Projects  Project Id  Memberships Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Upsert Project Membership Endpoint",
        "operationId": "upsert_project_membership_endpoint_api_projects__project_id__memberships_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectMembershipCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMembership"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/memberships/{user_id}": {
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project Membership Endpoint",
        "operationId": "delete_project_membership_endpoint_api_projects__project_id__memberships__user_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/export": {
      "get": {
        "tags": [
          "data-ops"
        ],
        "summary": "Export Project Endpoint",
        "operationId": "export_project_endpoint_api_projects__project_id__export_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectExportBundle-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/import": {
      "post": {
        "tags": [
          "data-ops"
        ],
        "summary": "Import Project Endpoint",
        "operationId": "import_project_endpoint_api_projects_import_post",
        "parameters": [
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/backup": {
      "post": {
        "tags": [
          "data-ops"
        ],
        "summary": "Backup Project Endpoint",
        "operationId": "backup_project_endpoint_api_projects__project_id__backup_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-nexus-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-User"
            }
          },
          {
            "name": "x-nexus-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Nexus-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectBackupResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_user_id": {
            "type": "string",
            "title": "Owner User Id",
            "default": "default-user"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Project"
      },
      "ProjectArchiveUpdate": {
        "properties": {
          "archived": {
            "type": "boolean",
            "title": "Archived"
          }
        },
        "type": "object",
        "required": [
          "archived"
        ],
        "title": "ProjectArchiveUpdate"
      },
      "ProjectBackupResult": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "backup_file": {
            "type": "string",
            "title": "Backup File"
          },
          "backup_path": {
            "type": "string",
            "title": "Backup Path"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "backup_file",
          "backup_path"
        ],
        "title": "ProjectBackupResult"
      },
      "ProjectCreate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectCreate"
      },
...
      "ProjectExportBundle-Output": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version",
            "default": "2026-03-15"
          },
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Exported At"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectSnapshot"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/ProjectMembershipSnapshot"
            },
            "type": "array",
            "title": "Memberships"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSnapshot"
            },
            "type": "array",
            "title": "Tasks"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/TaskDependencySnapshot"
            },
            "type": "array",
            "title": "Dependencies"
          },
          "context_entries": {
            "items": {
              "$ref": "#/components/schemas/ContextEntrySnapshot"
            },
            "type": "array",
            "title": "Context Entries"
          },
          "activity_events": {
            "items": {
              "$ref": "#/components/schemas/ActivityEventSnapshot"
            },
            "type": "array",
            "title": "Activity Events"
          }
        },
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectExportBundle"
      },
      "ProjectImportRequest": {
        "properties": {
          "bundle": {
            "$ref": "#/components/schemas/ProjectExportBundle-Input"
          },
          "target_project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Project Id"
          },
          "replace_existing": {
            "type": "boolean",
            "title": "Replace Existing",
            "default": false
          },
          "include_memberships": {
            "type": "boolean",
            "title": "Include Memberships",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "bundle"
        ],
        "title": "ProjectImportRequest"
      },
      "ProjectImportResult": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "replaced_existing": {
            "type": "boolean",
            "title": "Replaced Existing",
            "default": false
          },
          "imported_task_count": {
            "type": "integer",
            "title": "Imported Task Count",
            "default": 0
          },
          "imported_dependency_count": {
            "type": "integer",
            "title": "Imported Dependency Count",
            "default": 0
          },
          "imported_context_entry_count": {
            "type": "integer",
            "title": "Imported Context Entry Count",
            "default": 0
          },
          "imported_activity_event_count": {
            "type": "integer",
            "title": "Imported Activity Event Count",
            "default": 0
          },
          "imported_membership_count": {
            "type": "integer",
            "title": "Imported Membership Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "ProjectImportResult"
      },
      "ProjectMembership": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "member"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "user_id"
        ],
        "title": "ProjectMembership"
      },
      "ProjectMembershipCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ProjectMembershipCreate"
      },
      "ProjectMembershipSnapshot": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "member"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ProjectMembershipSnapshot"
      },
      "ProjectSnapshot": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_user_id": {
            "type": "string",
            "title": "Owner User Id",
            "default": "default-user"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectSnapshot"
      },
...
      "TaskSnapshot": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "todo"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "medium"
          },
          "labels": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TaskSnapshot"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "todo",
          "in_progress",
          "done"
        ],
        "title": "TaskStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}